+++ src/components/welcomeScreen/WelcomeScreen.scss
